+++ cligen.nim
@@ -16,16 +16,6 @@ proc dispatchId(name: string="", cmd: string="", rep: string=""): NimNode =
            elif cmd.len > 0: ident("dispatch" & cmd)  #XXX illegal chars
            else: ident("dispatch" & rep)
 
-proc toString(c: char): string =
-  ##Creates a string from char ``c``
-  result = newStringOfCap(1)
-  if c != '\0': result.add(c)
-
-proc toStrLitNode(n: NimNode): NimNode =
-  ##Creates a string literal node from a char literal NimNode
-  result = newNimNode(nnkStrLit)
-  result.strVal = toString(chr(n.intVal))
-
 proc toStrSeq(strSeqInitializer: NimNode): seq[string] =
   result = newSeq[string]()
   if strSeqInitializer.len > 1:
@@ -83,29 +73,31 @@ proc parseShorts(shorts: NimNode, proNm: auto, fpars: auto): Table[string,char]=
          lo != "version" and lo != "help" and lo != "help-syntax":
       error $proNm & " has no param matching `short` key \"" & lo & "\""
 
-proc dupBlock(fpars: NimNode, posIx: int, hlpCh: NimNode,
-              userSpec: Table[string, char]): Table[string, char] =
+proc dupBlock(fpars: NimNode, posIx: int, userSpec: Table[string, char]):
+     Table[string, char] =
   # Compute a table giving the short option for any long option, being
   # careful to only allow one such short option if the 1st letters of
   # two or more long options collide.
   result = initTable[string, char]()         # short option for param
+  var used: set[char] = {}                   # used shorts; bit vector ok
+  if "help" notin userSpec:
+    result["help"] = 'h'
+    used.incl('h')
   if "" in userSpec: return                  # Empty string key==>no short opts
-  var used: set[char]={ chr(hlpCh.intVal) }  # used shorts; bit vector ok
   for lo, sh in userSpec:
     result[lo] = sh
     used.incl(sh)
   for i in 1 ..< len(fpars):                 # [0] is proc, not desired here
     if i == posIx: continue                  # positionals get no option char
     let parNm = $(fpars[i][0])
-    if parNm.len == 1 and parNm[0] == chr(hlpCh.intVal):
-      error "Error: `" & parNm & "` collides with `shortHelp`.  Use another " &
-            "shortHelp='char' in `dispatch`."
+    if parNm.len == 1 and parNm[0] == result["help"]:
+      error "`"&parNm&"` collides with `short[\"help\"]`.  Change help short."
     let sh = parNm[0]                        # abbreviation is 1st character
     if sh notin used and parNm notin result: # still available
       result[parNm] = sh
       used.incl(sh)
-  let tmp = result
-  for k, v in tmp:
+  let tmp = result        #One might just put sh != '\0' above, but this lets
+  for k, v in tmp:        #CLI authors also block -h via short={"help": '\0'}.
     if v == '\0': result.del(k)
 
 proc collectComments(buf: var string, n: NimNode, depth: int = 0) =
@@ -152,17 +144,8 @@ proc posIxGet(positional: string, fpars: NimNode): int =
 proc newParam(id: string, rhs: NimNode): NimNode =
   return newNimNode(nnkExprEqExpr).add(ident(id), rhs)
 
-const helpTabOption*  = 0
-const helpTabType*    = 1
-const helpTabDefault* = 2
-const helpTabDescrip* = 3
-const helpTabColsDfl* = @[ helpTabOption, helpTabType,
-                           helpTabDefault, helpTabDescrip ]
-
-type Version* = tuple[longOpt: string, output: string]
-
-const dflUsage* = "${prelude}$command $args\n" &
-                  "$doc  Options(opt-arg sep :|=|spc):\n" & "$options"
+const clUse* = "$command $args\n$doc  Options(opt-arg sep :|=|spc):\n$options"
+const clUsage* = "Usage:\n  " & clUse
 
 type
   ClStatus* = enum clBadKey,                        ## Unknown long key
@@ -178,10 +161,41 @@ type
                    unparsedVal: string, ## Unparsed val ("" for missing)
                    message: string,     ## default error message
                    status: ClStatus]    ## Parse status for param
-
 const ClErrors* = { clBadKey, clBadVal, clNonOption, clMissing }
 const ClExit*   = { clHelpOnly, clVersionOnly }
 const ClNoCall* = ClErrors + ClExit
+
+type
+  ClHelpCol* = enum clOptKeys, clValType, clDflVal, clDescrip
+
+  ClCfg* = object  ## Settings tending to be program- or CLI-author-global
+    version*:     string
+    hTabCols*:    seq[ClHelpCol] ## selects columns to format
+    hTabRowSep*:  string         ## separates rows, e.g. "\n" double spaces
+    hTabColGap*:  int            ## number of spaces to separate cols by
+    hTabMinLast*: int            ## narrowest rightmost col no matter term width
+    hTabVal4req*: string         ## ``"REQUIRED"`` (or ``"NEEDED"``, etc.).
+    reqSep*:      bool           ## ``parseopt3.initOptParser`` parameter
+    sepChars*:    set[char]      ## ``parseopt3.initOptParser`` parameter
+    opChars*:     set[char]      ## ``parseopt3.initOptParser`` parameter
+
+{.push hint[GlobalVar]: off.}
+var clCfg* = ClCfg(
+  version:     "",
+  hTabCols:    @[ clOptKeys, clValType, clDflVal, clDescrip ],
+  hTabRowSep:  "",
+  hTabColGap:  2,
+  hTabMinLast: 16,
+  hTabVal4req: "REQUIRED",
+  reqSep:      false,
+  sepChars:    { '=', ':' },
+  opChars:     { '+', '-', '*', '/', '%', '@', ',', '.', '&',
+                 '|', '~', '^', '$', '#', '<', '>', '?'})
+{.pop.}
+
+proc toInts*(x: seq[ClHelpCol]): seq[int] =
+  ##Internal routine to convert help column enums to just ints for `alignTable`.
+  for e in x: result.add(int(e))
 
 proc contains*(x: openArray[ClParse], paramName: string): bool =
   ##Test if the ``seq`` updated via ``setByParse`` contains a parameter.
@@ -206,80 +220,54 @@ proc next*(x: openArray[ClParse], stati: set[ClStatus], start=0): int =
 
 include cligen/syntaxHelp
 
-macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
- help: typed = {}, short: typed = {}, usage: string=dflUsage,
- prelude="Usage:\n  ", echoResult: bool=false, requireSeparator: bool=false,
- sepChars={'=',':'},
- opChars={'+','-','*','/','%','@',',','.','&','|','~','^','$','#','<','>','?'},
- helpTabColumnGap: int=2, helpTabMinLast: int=16, helpTabRowSep: string="",
- helpTabColumns: seq[int] = helpTabColsDfl, stopWords: seq[string] = @[],
- positional: static string = AUTO, suppress: seq[string] = @[],
- shortHelp = 'h', implicitDefault: seq[string] = @[], mandatoryHelp="REQUIRED",
- mandatoryOverride: seq[string] = @[], version: Version=("",""),
- noAutoEcho: bool=false, dispatchName: string = "",
- mergeNames: seq[string] = @[], docs: ptr var seq[string]=nil,
- setByParse: ptr var seq[ClParse]=nil): untyped =
-  ##Generate command-line dispatcher for proc ``pro`` with extra help ``usage``.
-  ##Parameters without defaults in the proc become mandatory command arguments
-  ##while those with default values become command options.  Proc parameters
-  ##and option keys are normalized so that command users may spell multi-word
-  ##option keys flexibly as in ``--dry-Run | --dryrun``.  Each proc parameter
-  ##type must have in-scope ``argParse`` and ``argHelp`` procs (``argcvt.nim``
-  ##defines ``argParse/Help`` for many basic types, ``seq[T], set[T], ..``).
+macro dispatchGen*(pro: typed{nkSym}, cmdName: string="", doc: string="",
+  help: typed={}, short: typed={}, usage: string=clUsage, cf: ClCfg=clCfg,
+  echoResult=false, noAutoEcho=false, positional: static string=AUTO,
+  suppress: seq[string] = @[], implicitDefault: seq[string] = @[],
+  mandatoryOverride: seq[string] = @[], stopWords: seq[string] = @[],
+  dispatchName="", mergeNames: seq[string] = @[], docs: ptr var seq[string]=nil,
+  setByParse: ptr var seq[ClParse]=nil): untyped =
+  ##Generate command-line dispatcher for proc ``pro`` named ``dispatchName``
+  ##(defaulting to ``dispatchPro``) with generated help/syntax guided by
+  ##``cmdName``, ``doc``, and ``cf``.  Parameters with no explicit default in
+  ##the proc become mandatory command arguments while those with default values
+  ##become command options. Each proc parameter type needs in-scope ``argParse``
+  ##& ``argHelp`` procs.  ``cligen/argcvt`` defines them for basic types & basic
+  ##collections of them (``seq[T], set[T], ..``).
   ##
   ##``help`` is a ``{(paramNm, str)}`` of per-param help, eg. ``{"quiet": "be
   ##quiet"}``.  Often, only these help strings are needed for a decent CLI.
   ##
-  ##``short`` is a ``{(paramNm,char)}`` of per-param single-char option keys.
+  ##``short`` is a ``{(paramNm, char)}`` of per-param single-char option keys.
   ##
-  ##Since programs can return integer exit codes (often 1-byte) to OSes, if the
-  ##proc return is convertible to ``int`` that value is propagated unless
-  ##``echoResult`` is true.  However, if ``echoResult`` is true or if the
-  ##result is unconvertible and ``noAutoEcho`` is false then the generated
-  ##dispatcher echos the result of wrapped procs.  (Technically, dispatcher
-  ##callers like ``cligenQuit`` implement this behavior.)
+  ##``usage`` is a help template interpolating $command $args $doc $options.
   ##
-  ##If ``requireSeparator`` is true, both long & short options need an element
-  ##of ``sepChars`` before option values (if there are any). Any series of chars
-  ##in ``opChars`` may prefix an element of ``sepChars`` as in ``parseopt3``.
+  ##``cf`` controls whole program aspects of generated CLIs. See ``ClCfg`` docs.
   ##
-  ##``stopWords`` is a seq[string] of words beyond which ``-.*`` will no
-  ##longer signify an option (like the common sole ``--`` command argument).
-  ##
-  ##``helpTabColumnGap`` and ``helpTabMinLast`` control format parameters of the
-  ##options help table, and ``helpTabRowSep`` ("" by default) separates rows.
-  ##``helpTabColumns`` selects columns to format and is a seq of some subset of
-  ##``{ helpTabOption, helpTabType, helpTabDefault, helpTabDescrip }``, though
-  ##only the final column in a help table row auto-word-wraps.
+  ##Default exit protocol is: quit(int(result)) or (echo $result or discard;
+  ##quit(0)) depending on what compiles.  True ``echoResult`` forces echo while
+  ##``noAutoEcho`` blocks it (=>int()||discard).  Technically, ``cligenQuit``
+  ##implements this behavior.
   ##
   ##By default, ``cligen`` maps the first non-defaulted ``seq[]`` proc parameter
   ##to any non-option/positional command args.  ``positional`` selects another.
   ##Set ``positional`` to the empty string (``""``) to disable this entirely.
   ##
-  ##``suppress`` is a list of formal parameter names to NOT include in the
-  ##parsing/assigning system.  Such names are effectively pinned to whatever
-  ##their default values are.
-  ##
-  ##``shortHelp`` is a char to use for a short option key analogue of --help.
+  ##``suppress`` is a list of formal parameter names to exclude from the parse-
+  ##assign system.  Such names are effectively pinned to their default values.
   ##
   ##``implicitDefault`` is a list of formal parameter names allowed to default
-  ##to the Nim default value for a type, rather than becoming mandatory, even
-  ##when they are missing an explicit initializer. ``mandatoryHelp`` is how the
-  ##default value appears in help messages for mandatory parameters.
+  ##to the Nim default value for a type, rather than becoming mandatory, when
+  ##they lack an explicit initializer.
   ##
-  ##``mandatoryHelp`` is the default value string in help tables for required
-  ##parameters.  ``mandatoryOverride`` is a list of strings indicating parameter
-  ##names which override mandatory-ness of anything else.
+  ##``mandatoryOverride`` is a list of formal parameter names which, if set by a
+  ##CLI user, override the erroring out if any mandatory settings are missing.
   ##
-  ##``version`` is a ``Version`` 2-tuple ``(longOpt for vrsn`, `vrsn string)``
-  ##which defines how a CLI user may dump the version of a program.  If you
-  ##want to provide a short option, add a ``"version":'v'`` entry to ``short``.
+  ##``stopWords`` is a ``seq[string]`` of words beyond which ``-.*`` no longer
+  ##signifies an option (like the common sole ``--`` command argument).
   ##
-  ##``dispatchName`` is the name of a generated dispatcher, defaulting to simply
-  ##``"dispatchpro"`` where ``pro`` is the name of the proc being wrapped.
-  ##
-  ##``mergeNames`` gives the ``cmdNames`` param passed to ``mergeParams``,
-  ##which defaults to ``@[cmdName]`` if ``mergeNames`` is ``@[]``.
+  ##``mergeNames`` gives the ``cmdNames`` param passed to ``mergeParams``, which
+  ##defaults to ``@[cmdName]`` if ``mergeNames`` is ``@[]``.
   ##
   ##``docs`` is ``addr(some var seq[string])`` to which to append each main doc
   ##comment or its replacement doc=text.  Default of ``nil`` means do nothing.
@@ -295,9 +283,6 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
   ##this ``var seq`` needing to be declared before such procs for such access is
   ##ok.  Ideally, keep important functionality Nim-callable.  ``setByParse`` may
   ##also be useful combined with the ``parseOnly`` arg of generated dispatchers.
-
-  #XXX quote-do fails to access macro args in sub-scopes. So `help`, `cmdName`..
-  #XXX need either to be used at top-level or assigned in a shadow local.
   let impl = pro.getImpl
   if impl == nil: error "getImpl(" & $pro & ") returned nil."
   let fpars = formalParams(impl, toStrSeq(suppress))
@@ -305,12 +290,15 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
   if cmtDoc.len == 0:                   # allow caller to override commentDoc
     collectComments(cmtDoc, impl)
     cmtDoc = strip(cmtDoc)
+  let cf = cf #sub-scope quote-do's cannot access macro args w/o shadow locals.
   let proNm = $pro                      # Name of wrapped proc
   let cName = if len($cmdName) == 0: proNm else: $cmdName
   let disNm = dispatchId($dispatchName, cName, proNm) # Name of dispatch wrapper
   let helps = parseHelps(help, proNm, fpars)
   let posIx = posIxGet(positional, fpars) #param slot for positional cmd args|-1
-  let shOpt = dupBlock(fpars, posIx, shortHelp, parseShorts(short,proNm,fpars))
+  var shortTab = parseShorts(short, proNm, fpars)
+  let shOpt = dupBlock(fpars, posIx, shortTab)
+  let shortH = shOpt["help"]
   var spars = copyNimTree(fpars)        # Create shadow/safe suffixed params.
   var dpars = copyNimTree(fpars)        # Create default suffixed params.
   var mandatory = newSeq[int]()         # At the same time, build metadata on..
@@ -338,9 +326,7 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
   let keyCountId = ident("keyCount")    # positional arg number
   let usageId = ident("usage")          # gen proc parameter
   let cmdLineId = ident("cmdline")      # gen proc parameter
-  let vsnOpt = $version[0]              # Need string lits here for CL parse
-  let vsnSh = if vsnOpt in shOpt: $shOpt[vsnOpt] else: ""
-  let vsnStr = version[1]               # value must just work in stdout.write
+  let vsnSh = if "version" in shOpt: $shOpt["version"] else: "\0"
   let prefixId = ident("prefix")        # local help prefix param
   let pId = ident("p")                  # local OptParser result handle
   let allId = ident("allParams")        # local list of all parameters
@@ -350,12 +336,7 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
   let apId = ident("ap")                # ArgcvtParams
   var callIt = newNimNode(nnkCall)      # call of wrapped proc in genproc
   callIt.add(pro)
-  let htColGap = helpTabColumnGap
-  let htMinLst = helpTabMinLast
-  let htRowSep = helpTabRowSep
-  let htCols   = helpTabColumns
-  let prlude   = prelude; let mandHelp = mandatoryHelp
-  let shortHlp = shortHelp
+  let shortHlp = newStrLitNode($shortH)
   let setByParseId = ident("setByP")    # parse recording var seq
   let setByParseP = setByParse
 
@@ -364,8 +345,8 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
     let tabId = ident("tab")            # local help table var
     result.add(quote do:
       var `apId`: ArgcvtParams
-      `apId`.mand = `mandHelp`
-      let shortH = $(`shortHlp`)
+      `apId`.val4req = `cf`.hTabVal4req
+      let shortH = `shortHlp`
       var `allId`: seq[string] = @[ "help", "help-syntax" ]
       var `cbId`: CritBitTree[string]
       `cbId`.incl(optionNormalize("help"), "help")
@@ -378,12 +359,11 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
       `apId`.shortNoVal = { shortH[0] }               # argHelp(bool) updates
       `apId`.longNoVal = @[ "help", "help-syntax" ]   # argHelp(bool) appends
       let `setByParseId`: ptr seq[ClParse] = `setByParseP`)
-    if vsnOpt.len > 0:
-      result.add(quote do:
-       var versionDflt = false
-       `apId`.parNm = `vsnOpt`; `apId`.parSh = `vsnSh`; `apId`.parReq = 0
-       `apId`.parRend = `vsnOpt`
-       `tabId`.add(argHelp(versionDflt, `apId`) & "print version"))
+    result.add(quote do:
+      if `cf`.version.len > 0:
+        `apId`.parNm = "version"; `apId`.parSh = `vsnSh`
+        `apId`.parReq = 0; `apId`.parRend = `apId`.parNm
+        `tabId`.add(argHelp(false, `apId`) & "print version"))
     let argStart = if mandatory.len > 0: "[required&optional-params]" else:
                                          "[optional-params]"
     let posHelp = if posIx != -1:
@@ -402,7 +382,7 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
       callIt.add(newNimNode(nnkExprEqExpr).add(idef[0], sdef[0])) #Add to call
       if i != posIx:
         let parNm = $idef[0]
-        let sh = toString(shOpt.getOrDefault(parNm))      #Add to perPar helpTab
+        let sh = $shOpt.getOrDefault(parNm)      #Add to perPar helpTab
         let defVal = sdef[0]
         let hlp =
           if parNm in helps:
@@ -415,19 +395,19 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
          `apId`.parRend = helpCase(`parNm`, clLongOpt)
          let descr = getDescription(`defVal`, `parNm`, `hlp`)
          `tabId`.add(argHelp(`defVal`, `apId`) & descr)
-         if `apId`.parReq != 0: `tabId`[^1][2] = `apId`.mand
+         if `apId`.parReq != 0: `tabId`[^1][2] = `apId`.val4req
          `cbId`.incl(optionNormalize(`parNm`), `apId`.parRend)
          `allId`.add(helpCase(`parNm`, clLongOpt)))
         if isReq:
           result.add(quote do: `mandId`.add(`parNm`))
     result.add(quote do:                  # build one large help string
       let indentDoc = addPrefix(`prefixId`, wrap(`prefixId`, `cmtDoc`))
-      `apId`.help = `usageId` % [ "prelude", `prlude`, "doc", indentDoc,
-                     "command", `cName`, "args", `args`, "options",
-                     addPrefix(`prefixId` & "  ",
-                               alignTable(`tabId`, 2*len(`prefixId`) + 2,
-                                          `htColGap`, `htMinLst`, `htRowSep`,
-                                          `htCols`))]
+      `apId`.help = `usageId` % [ "doc",indentDoc, "command",`cName`,
+                                  "args",`args`, "options",
+                  addPrefix(`prefixId` & "  ",
+                            alignTable(`tabId`, 2*len(`prefixId`) + 2,
+                                       `cf`.hTabColGap, `cf`.hTabMinLast,
+                                       `cf`.hTabRowSep, toInts(`cf`.hTabCols)))]
       if `apId`.help.len > 0 and `apId`.help[^1] != '\n':   #ensure newline @end
         `apId`.help &= "\n"
       if len(`prefixId`) > 0:             # to indent help in a multicmd context
@@ -437,7 +417,7 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
     result = newNimNode(nnkCaseStmt).add(quote do:
       if p.kind == cmdLongOption: lengthen(`cbId`, `pId`.key) else: `pId`.key)
     result.add(newNimNode(nnkOfBranch).add(
-      newStrLitNode("help"), toStrLitNode(shortHlp)).add(
+      newStrLitNode("help"), shortHlp).add(
         quote do:
           if cast[pointer](`setByParseId`) != nil:
             `setByParseId`[].add(("help", "", `apId`.help, clHelpOnly))
@@ -452,24 +432,14 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
             return                            #Do not try to keep parsing
           else:
             stdout.write(syntaxHelp); raise newException(HelpOnly, "")))
-    if vsnOpt.len > 0:
-      if vsnOpt in shOpt:                     #There is also a short version tag
-        result.add(newNimNode(nnkOfBranch).add(
-          newStrLitNode(vsnOpt), newStrLitNode(vsnSh)).add(
-            quote do:
-              if cast[pointer](`setByParseId`) != nil:
-                `setByParseId`[].add((`vsnOpt`, "", `vsnStr`, clVersionOnly))
-                return                        #Do not try to keep parsing
-              else:
-                stdout.write(`vsnStr`,"\n");raise newException(VersionOnly,"")))
-      else:                                   #There is only a long version tag
-        result.add(newNimNode(nnkOfBranch).add(newStrLitNode(vsnOpt)).add(
-            quote do:
-              if cast[pointer](`setByParseId`) != nil:
-                `setByParseId`[].add((`vsnOpt`, "", `vsnStr`, clVersionOnly))
-                return                        #Do not try to keep parsing
-              else:
-                stdout.write(`vsnStr`,"\n");raise newException(VersionOnly,"")))
+    result.add(newNimNode(nnkOfBranch).add(
+      newStrLitNode("version"), newStrLitNode(vsnSh)).add(
+        quote do:
+          if cast[pointer](`setByParseId`) != nil:
+            `setByParseId`[].add(("version", "", `cf`.version, clVersionOnly))
+            return                        #Do not try to keep parsing
+          else:
+            stdout.write(`cf`.version,"\n");raise newException(VersionOnly,"")))
     for i in 1 ..< len(fpars):                # build per-param case clauses
       if i == posIx: continue                 # skip variable len positionals
       let parNm  = $fpars[i][0]
@@ -595,7 +565,7 @@ macro dispatchGen*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
         var `posNoId` = 0
         var `keyCountId` = initCountTable[string]()
         var `pId` = initOptParser(args, `apId`.shortNoVal, `apId`.longNoVal,
-                                  `requireSeparator`, `sepChars`, `opChars`,
+                                  `cf`.reqSep, `cf`.sepChars, `cf`.opChars,
                                   `stopWords`)
         while true:
           next(`pId`)
@@ -666,26 +636,16 @@ macro cligenQuitAux*(cmdLine:seq[string], dispatchName: string, cmdName: string,
   quote do: cligenQuit(`disNm`(mergeParams(`mergeNms`, `cmdLine`)),
                        `echoResult`, `noAutoEcho`)
 
-template dispatch*(pro: typed{nkSym}, cmdName: string = "", doc: string = "",
- help: typed = {}, short: typed = {}, usage: string=dflUsage,
- prelude="Usage:\n  ", echoResult: bool=false, requireSeparator: bool=false,
- sepChars={'=',':'},
- opChars={'+','-','*','/','%','@',',','.','&','|','~','^','$','#','<','>','?'},
- helpTabColumnGap: int=2, helpTabMinLast: int=16, helpTabRowSep: string="",
- helpTabColumns = helpTabColsDfl, stopWords: seq[string] = @[],
- positional = AUTO, suppress: seq[string] = @[],
- shortHelp = 'h', implicitDefault: seq[string] = @[], mandatoryHelp="REQUIRED",
- mandatoryOverride: seq[string] = @[], version: Version=("",""),
- noAutoEcho: bool=false, dispatchName: string = "",
- mergeNames: seq[string] = @[]): untyped =
+template dispatch*(pro: typed{nkSym}, cmdName="", doc="", help: typed={},
+ short:typed={},usage=clUsage, cf:ClCfg=clCfg,echoResult=false,noAutoEcho=false,
+ positional=AUTO, suppress:seq[string] = @[], implicitDefault:seq[string] = @[],
+ mandatoryOverride: seq[string] = @[], dispatchName="",
+ mergeNames: seq[string] = @[], stopWords: seq[string] = @[]): untyped =
   ## A convenience wrapper to both generate a command-line dispatcher and then
-  ## call the dispatcher & exit; Usage is the same as the ``dispatchGen`` macro.
-  dispatchGen(
-    pro, cmdName, doc, help, short, usage, prelude, echoResult,
-      requireSeparator, sepChars, opChars, helpTabColumnGap, helpTabMinLast,
-      helpTabRowSep, helpTabColumns, stopWords, positional, suppress, shortHelp,
-      implicitDefault, mandatoryHelp, mandatoryOverride, version, noAutoEcho,
-      dispatchName)
+  ## call the dispatcher & exit; Params are same as the ``dispatchGen`` macro.
+  dispatchGen(pro, cmdName, doc, help, short, usage, cf, echoResult, noAutoEcho,
+              positional, suppress, implicitDefault, mandatoryOverride,
+              stopWords, dispatchName, mergeNames)
   cligenQuitAux(os.commandLineParams(), dispatchName, cmdName, pro, echoResult,
                 noAutoEcho)
 
@@ -723,12 +683,8 @@ proc subCmdNoAutoEc(node: NimNode): bool =
     if child.kind == nnkExprEqExpr and eqIdent(child[0], "noAutoEcho"):
       return true
 
-{.push hint[GlobalVar]: off.}
-var cligenVersion* = ""
-{.pop.}
-
 proc subCmdUsage(node: NimNode): string =
-  result = dflUsage
+  result = clUse
   for child in node:
     if child.kind == nnkExprEqExpr and eqIdent(child[0], "usage"):
       return child[1].strVal
@@ -765,9 +721,9 @@ $1 --help-syntax gives general cligen syntax help.
 Run "$1 {help CMD|CMD --help}" to see help for just CMD.
 Run "$1 help" to get *comprehensive* help.$3""" % [ srcBase,
   addPrefix("  ", alignTable(pairs, prefixLen=2)),
-  (if cligenVersion.len > 0: "\nTop-level --version also available" else: "") ]
+  (if clCfg.version.len > 0: "\nTop-level --version also available" else: "") ]
 
-proc srcBaseName*(n: NimNode): NimNode =
+proc srcBaseName(n: NimNode): NimNode =
   let fileParen = lineinfo(n)      # Infer multi-cmd name from lineinfo
   let slash = if rfind(fileParen, "/") < 0: 0 else: rfind(fileParen, "/") + 1
   let paren = rfind(fileParen, ".nim(") - 1
@@ -813,8 +769,8 @@ macro dispatchMultiGen*(procBkts: varargs[untyped]): untyped =
     let sCmdNm = subCmdName(p)
     var c = newCall("dispatchGen")
     copyChildrenTo(p, c)
-    if not c.paramPresent("prelude"):
-      c.add(newParam("prelude", newStrLitNode("")))
+    if not c.paramPresent("usage"):
+      c.add(newParam("usage", newStrLitNode(clUse)))
     if not c.paramPresent("mergeNames"):
       c.add(newParam("mergeNames", quote do: @[ `srcBase`, `sCmdNm` ]))
     if not c.paramPresent("docs"):
@@ -879,9 +835,7 @@ macro dispatchMultiGen*(procBkts: varargs[untyped]): untyped =
       unknownSubcommand(`arg0Id`, `subCmdsId`)))
   result.add(quote do:
     `multiNmsId`.add("dispatch" & `prefix`)
-    proc `multiId`(`cmdLineId`: seq[string],
-                   `usageId`=dflUsage,
-                   `prefixId`="  ") =
+    proc `multiId`(`cmdLineId`: seq[string], `usageId`=clUse, `prefixId`="  ") =
       {.push hint[XDeclaredButNotUsed]: off.}
       let n = `cmdLineId`.len
       let `arg0Id` = if n > 0: `subMchsId`.lengthen `cmdLineId`[0] else: ""
@@ -905,15 +859,13 @@ macro dispatchMultiDG*(procBkts: varargs[untyped]): untyped =
     result[^1].add(newParam("stopWords", subCmdsId))
   if not result[^1][0].paramPresent("dispatchName"):
     result[^1].add(newParam("dispatchName", newStrLitNode(prefix & "Subs")))
-  if not result[^1][0].paramPresent("version"):
-    result[^1].add(newParam("version", quote do: ("version", cligenVersion)))
   if not result[^1][0].paramPresent("suppress"):
     result[^1].add(newParam("suppress", quote do: @[ "usage", "prefix" ]))
   let srcBase = srcBaseName(procBrackets)
   let subDocsId = ident(prefix & "SubDocs")
   if not result[^1][0].paramPresent("usage"):
     result[^1].add(newParam("usage", quote do:
-      "${prelude}" & topLevelHelp(`srcBase`, `subCmdsId`, `subDocsId`)))
+      "Usage:\n  " & topLevelHelp(`srcBase`, `subCmdsId`, `subDocsId`)))
   when defined(printDispatchDG): echo repr(result)  # maybe print gen code
 
 macro dispatchMulti*(procBrackets: varargs[untyped]): untyped =
@@ -962,3 +914,5 @@ proc mergeParams*(cmdNames: seq[string],
   ##context, ``cmdNames[0]`` is the ``cmdName`` while in a ``dispatchMulti
   ##``context it is ``@[ <mainCommand>, <subCommand> ]``.
   cmdLine
+
+include cligen/oldAPI   #TEMPORARY SUPPORT FOR THE OLD CALLING CONVENTION