+++ test/EchoResult.nim
@@ -3,4 +3,4 @@ when NimVersion <= "0.19.8":
   import editDistance
 else:
   import std/editDistance
-dispatch(editDistanceAscii, echoResult=true, requireSeparator=true)
+dispatch(editDistanceAscii, echoResult=true)