+++ test/MultMultMult.nim
@@ -30,7 +30,7 @@ proc yikes(hooves=4, races=9, verb=false, names: seq[string]): string =
 when isMainModule:
   import cligen
   include cligen/mergeCfgEnv
-  cligenVersion = "0.0.1"
+  clCfg.version = "0.0.1"
   dispatchMultiGen([ "cobbler" ],
                    [ yikes,
                      mergeNames = @["MultiMulti", "apple", "cobbler" ]])